spacemanager: Fix S2 usecase.CheckGarbageSpaceCollector regression
S2 insists that a space expires immediately as soon as its lifetime
has passed. This conflicts with the lazy expiration implemented by
the spacemanager. The SRM used to have a workaround for this, but
that workaround was recently removed.

This patch adds a check to space manager that will report the state
as expired once the lifetime has been exceeded.

Target: trunk
Request: 2.7
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6227/