(2.7) dcache-webadmin: remove autorefresh on pages with ListViews
The introduction of autorefresh (the addition of the AjaxSelfUpdatingTimerBehavior) to the Forms on pages containing ListViews that also have javascript sort and search functions associated with their headers has introduced a bug which manifests itself as the disappearance, after refresh, of the search boxes in the header and the inability to sort by clicking on the column title.

After some investigation, it seems that this has to do with the way Wicket handles Ajax requests.

I have not found a way to get this cocktail of javascript, Wicket and Ajax to play correctly for the "repeater" (ListView) classes.  The Alarms page uses a different component, the DataTable, which does not have this issue.  Changing all the ListViews that need to be refreshed into DataTables would be rather invasive, however, so for 2.7 I have decided simply to revert to no autorefresh for these pages, making 2.7 essentially equivalent to 2.6.

I plan to make the component substitution for DataTables in master.

Testing: Redeployed instance behaves as 2.6 (the refreshed data must be obtained by clicking on the navigation link or reloading the page).

Target: 2.7
Patch: http://rb.dcache.org/r/6230/
Require-notes: no
Require-book: no
Acked-by: Gerd