billing: Add date prefix to output of search command
Our default billing format does not include the year in the billing
file, and it is unclear whether the order in the billing file is
month.year or year.month. This patch adds the full date as a prefix
to the output when searching for billing entries. The date is
formatted using the system local.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6251/