gPlazma-ldap: fix non-serializable Set problem
reverseMap returned a automatically Scala to Java converted Set class
that did not implement Serializable.  This caused an
NotSerializableException when the object should be send in a cell
message.  This patch adds a couple of manual conversions instead of
automatic ones, ensuring the returned Set is Serializable.

Ticket:
Acked-by:
Target: trunk
Require-book: no
Require-notes: no