cells: Move dCache's cell interfaces into cells
These interfaces were originally placed outside of cells
as cells was a closed product. Now that cells is part of
dcache, there is no reason not to move those interfaces
to the proper place.

This allows the LoginBrokerHandler utility class to be
moved to the same package on the LoginManager it is coupled
to. A future patch will update LoginManager to use that
utility class and thus avoid the code duplication with an
internal class.

This patch only moves classes. Besides deleting the unused
ThreadFactoryAware, there are no code changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6285/