srm: refactor part of Scheduler to use guava Service
This patch takes the code that tries to keep the executors busy and
makes a Service out of it, based on AbstractExecutionThreadService.
The core logic remains unchanged, but it is collected into a
class.

The goal is to refactor the different sub-components that Scheduler
uses into Service classes and use a ServiceManager to manage their
lifecycle.

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/6313/
Acked-by: Gerd Behrmann