poolmanager: Add named error codes for no-pool-configured and no-pool-online
Improves error messages when poolmanager is unable to find pools for a
transfer. In particular the 'no write pool configured' and 'no write pool
available' messages are simplified when sent to the FTP client, and the log
messages no longer use the legacy StorageInfo#toString method (instead only the
information that is actually involved in pool selection is logged).

Having two different error codes is important as one error is transient and the
other is not. Read pool selection currently uses some additional error codes
which should probably be changed to one of the two new values, but I leave that
for a future patch.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6358/