gplazma: ldap: make user home and root configurable
The LDAP plugin has hard coded values for the home and root directory as
they are presented to the user. The root directory is set to / and the
home directory is set to the home directory as stored in the LDAP
server. For operating in "cloud" mode this is not what we want, because
we want every user to see his home directory as the root directory.

This patch makes the behaviour configurable by introducing two new
properties gplazma.ldap.home-dir and gplazma.ldap.root-dir that can be
used to configure dcache to rather run in "cloud"-mode than in the usual
way. In those two properties it is now possible to use keywords (e.g.,
"%homeDirectory%" that will be substituted by the corresponding
attribute value in LDAP. For example %homeDirectory% will be replaced by
the value of attribute "homeDirectory" (in the peopleTree).

Acked-by: Paul
Target: trunk
Require-book: yes
Require-notes: yes