chimera: refactor PGET into abstract node
Given the fact that we would like to optimize the size of the file handles on the ".(get)" operations by not including the arguments, it now makes no sense to have the FsInode_PGET represent any concrete operation. Instead, we would like a hierarchy of get operations which have the basic read mechanism in common.

This patch represents the first of a two part refactoring.  It does the following:

1.  Changes FsInode_PGET to an abstract node.
2.  Implements the degenerate case (cursor) as a separate subclass FsInode_PCUR.
3.  Changes FsInode_CRC (checksum) to be a subclass of PGET.
4.  Adjusts the FsInodeType to reflect these changes.
5.  Modifies the JdcbFs class accordingly.

Testing: Works on both nfs 3 and nfs v4.1.

Target: master
Request: 2.7
Patch: http://rb.dcache.org/r/6407
Require-book: no
Require-notes: no
Acked-by: Tigran

Changes are invisible to user.