chimera: refactor PGET, part 2: eliminate caching of file locality node
Use of the Guava cache to maintain state on the FsInode used to convey locality information is unnecessary if the operation subclasses the abstract PGET operation.

The slight disadvantage is that a call to the PoolManager will be made each time stat() is called by the client; the advantage is that the server is not maintaining state on client operations, which does not really fit the NFS protocol and was something of a hack.

This patch completes the refactoring begun in 6407.  It changes the FsInode_PLOC to subclass PGET, removes the caching mechanism from the DCacheAware fs interface, and adds one method the FileSystem API (getFileLocality).

Testing: Tested successfully on nfs3 and nfs v4.1.

Target: master
Request: 2.7
Patch: http://rb.dcache.org/r/6409
Require-book: no
Require-notes: no
Depends-on: http://rb.dcache.org/r/6407
Acked-by: Tigran

Changes are invisible to the user.
(cherry picked from commit cf963c4a6a78cc92d27209512be98fdc047e6b33)

Signed-off-by: alrossi <arossi@otfrid.fnal.gov>