spacemanager: Limit ls files command to transient states by default
Defaulting to listing all files in stored state is a bit dangerous as it is
guaranteed to consume lots of memory on most production systems.

This patch changes the defaults to only list files in allocated and
transferring states. These are likely the files an admin would be interested
in, e.g. when wanting to release a lock manually.

Also fixed some help text.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6484/