spacemanager: Fix slow SQL query
On small tables the two queries generate the same plan, but when
testing it on a clone of our production system, the queries produce
significantly different results (despite being equivalent when
the columns involved are marked not null).

The new query produces the same query plan as a left outer join
query, so there is no reason to use an explicit join (which would
be difficult to get into the criterion api).

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/6487/