utils: move Args class into commons
An old copy of the Args class exists in srm-common package.  This patch
moves the Args class from cells to dcache-common so it may be used by
SRM and dCache code alike.

The patch removes Serializable from Args.  This is because we currently don't
use Args in Messages and want to discourage any future use.

The patch also updates the import section of the files affected to follow
the agreed ordering.

Target: master
Requres-book: no
Requires-notes: no
Patch: http://rb.dcache.org/r/6506/
Acked-by: Gerd Behrmann