Restore compatibility with 2.6 pools
CHANGE_TIME is not available in 2.6 code and the enum can thus
not be deserialized. This is obviously a flaw in the design of
the FileAttributes class making it difficult to add new
attributes, but for now this patch restored pool compatibility
by not including that attribute by default.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6510/