spacemanager: Improve CLI errors
In case of lookups for specific entities, the EmptyResultDataAccessExcption
message contains enough information to describe the error. This patch avoids
that a lengthy causal chain of exceptions are printed to the shell for something
that must be considered a common error.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6513/