pool: Bind FTP data channel to single interface
The FTP protocol only allows a sinle IP address to be returned to the client,
yet the mover listened on the wildcard address.  This has lead to surprising
results, as an application can by a socket to the wildcard address even when
another socket is already bound to the same port on a specific interface. This
could lead to situations in which the client would connect to the other socket
and not the socket created by the mover (e.g. the admin door that listens in
the same port range).

The patch addresses the problem by only binding the server socket to the
address we would return to the client. The patch changes the behaviour by
removing one DNS reverse lookup and DNS forward lookup: Previously the pool
would return the canonical name of the interface and the door would resolve
that back to an IP address before returning the address to the client. Now the
pool send an address to the door and the door just passes on that address. In
theory this could lead to changes in which pool interface is reported. For this
reason I do not request this patch to be merged to stable branches.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6593/
(cherry picked from commit 2dd9d0e7e27cef2c31df37fd9c80474b25c36d82)